refactor: split trade_order_validator into context + rules (#3427)
Phase 3 step 11: extract TradeOrderValidationContext, the rejection-reason
codes, and tradeOrderValidationContextFromGame into a new
trade_order_validation_context.dart, leaving trade_order_validator.dart with
only the TradeOrderValidator rule set. Both are re-exported from the
colonizethis_economy barrel so callers are unaffected. Brings
trade_order_validator.dart from 324 to 208 lines (under the 300 ceiling).

Refs #3427
