+++ packages/colonizethis_economy/lib/colonizethis_economy.dart
@@ -24,6 +24,7 @@ export 'src/economy/world_market/purchased_tile_index.dart';
 export 'src/economy/world_market/purchased_tile_riches.dart';
 export 'src/economy/world_market/sellable_quantity.dart';
 export 'src/economy/world_market/trade_order_suggester.dart';
+export 'src/economy/world_market/trade_order_validation_context.dart';
 export 'src/economy/world_market/trade_order_validator.dart';
 export 'src/economy/world_market/treasury_bid_budget.dart';
 export 'src/validation/order_validation_result.dart';
